common code file
