+++ gravity-restaking/src/lib.rs
@@ -3,6 +3,7 @@
 multiversx_sc::imports!();
 
 pub mod call_delegation;
+pub mod common_actions;
 pub mod sovereign;
 pub mod token_whitelist;
 pub mod unique_payments;
@@ -16,6 +17,7 @@ pub trait GravityRestaking:
     + user::UserModule
     + validator::ValidatorModule
     + sovereign::SovereignModule
+    + common_actions::CommonActionsModule
     + utils::UtilsModule
 {
     #[init]
